fix: tooltip 관련 파일명 변경
- floating-label -> tooltip
- floating-label-token -> tooltip-tag
