--- apps/client/.env.example
@@ -1,2 +1,0 @@
-VITE_BASE_URL= 'API 기본 URL'
-VITE_SENTRY_DSN='센트리DSN'
