--- packages/cds-ui/src/components/floating-label-token/floating-label-token.tsx
@@ -1,15 +1,0 @@
-import * as styles from './floating-label-token.css';
-
-interface FloatingLabelTokenProps {
-  children: string;
-}
-
-const FloatingLabelToken = ({ children }: FloatingLabelTokenProps) => {
-  return (
-    <button type="button" className={styles.container}>
-      {children}
-    </button>
-  );
-};
-
-export default FloatingLabelToken;
