+++ packages/cds-ui/src/components/index.ts
@@ -5,8 +5,6 @@ export { default as ConfirmModal } from './confirm-modal/confirm-modal';
 export { default as DetailModal } from './detail-modal/detail-modal';
 export { default as File } from './file/file';
 export { default as FloatingButton } from './floating-button/floating-button';
-export { default as FloatingLabel } from './floating-label/floating-label';
-export { default as FloatingLabelToken } from './floating-label-token/floating-label-token';
 export { default as FloatingMenu } from './floating-menu/floating-menu';
 export { default as ImageContainer } from './image-container/image-container';
 export { default as Label } from './label/label';
@@ -21,3 +19,5 @@ export { default as SideBarProfile } from './sidebar-profile/sidebar-profile';
 export { default as TextContent } from './text-content/text-content';
 export { default as Title } from './title/title';
 export { default as Toggle } from './toggle/toggle';
+export { default as Tooltip } from './tooltip/tooltip';
+export { default as TooltipTag } from './tooltip-tag/tooltip-tag';
